2.7 10/04 Версия 1
