+++ .gitignore
@@ -12,3 +12,5 @@ Thumbs.db
 node_modules/
 
 # Файлы сборки (допишите самостоятельно)
+./public/bundle.js
+./public/bundle.js.map
