+++ public/index.html
@@ -33,7 +33,7 @@
     <!-- Количество фильмов -->
   </section>
 </footer>
-<script src="../src/main.js"></script>
+<script src="./bundle.js"></script>
 </body>
 
 </html>
