+++ src/components/createBtnLoadMore.js
@@ -1,0 +1,7 @@
+const createBtnLoadMore = () => {
+  return (`
+    <button class="films-list__show-more">Show more</button>
+  `);
+};
+
+export {createBtnLoadMore};
