+++ src/components/createFilmsListContainer.js
@@ -1,0 +1,9 @@
+const createFilmsListContainer = () => {
+  return (`
+      <div class="films-list__container">
+
+      </div>
+`);
+};
+
+export {createFilmsListContainer};
