+++ src/components/createFilmsListExtraSection.js
@@ -1,0 +1,9 @@
+const createFilmsListExtraSection = () => {
+  return (`
+      <section class="films-list--extra">
+
+      </section>
+`);
+};
+
+export {createFilmsListExtraSection};
