+++ src/components/createFooterStatistics.js
@@ -1,0 +1,7 @@
+const createFooterStatistics = () => {
+  return (`
+  <p>130 291 movies inside</p>
+  `);
+};
+
+export {createFooterStatistics};
