+++ src/components/createHeader.js
@@ -1,0 +1,5 @@
+const createHeader = () => {
+  return (`<h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>`);
+};
+
+export {createHeader};
