+++ src/components/createHeaderMostCommented.js
@@ -1,0 +1,5 @@
+const createHeaderMostCommented = () => {
+  return (`<h2 class="films-list__title">Most commented</h2>`);
+};
+
+export {createHeaderMostCommented};
