+++ src/components/createHeaderTopRated.js
@@ -1,0 +1,5 @@
+const createHeaderTopRated = () => {
+  return (`<h2 class="films-list__title">Top rated</h2>`);
+};
+
+export {createHeaderTopRated};
