+++ src/components/createSectionFilms.js
@@ -1,0 +1,11 @@
+const createSectionFilms = () => {
+  return (`
+   <section class="films">
+    <section class="films-list">
+
+    </section>
+  </section>
+  `);
+};
+
+export {createSectionFilms};
